feat: move rfid input from +page.svelte to LoginForm.svelte
