+++ src/routes/+page.svelte
@@ -6,51 +6,11 @@
 
     export let form;
 
-    onMount(() => {
-        let rfidInput = document.getElementById('rfidInput'); //take the rfid input
-
-        const handleClickOutside = (e: any) => {
-            const target = e.target;    //take the target of the event
-            const focusedElement = document.activeElement;  //this is the active element being focused
-
-            // if the current element is rfidInput and you are not clicking an input or a button, do not remove the focus
-            if (focusedElement === rfidInput && !(target instanceof HTMLInputElement || target instanceof HTMLButtonElement)) {
-                e.preventDefault();
-            }
-
-            // if the current element is not the rfidInput and you click anything else, it should focus the rfid input
-            else if(focusedElement !== rfidInput && !(target instanceof HTMLInputElement || target instanceof HTMLButtonElement)) {
-
-                //need to add a delay because of svelte
-                setTimeout(() => {
-                    rfidInput?.focus()
-                }, 10)
-            }  
-        };
-        
-        const handleKeyDown = (e: any) => {
-            const focusedElement = document.activeElement;  //this is the active element being focused
-            if (rfidInput === focusedElement){
-                console.log()
-                e.preventDefault();
-            }
-        }
-
-        // when mouse is clicked, call handle click outside
-        document.addEventListener('mousedown', handleClickOutside);
-        document.addEventListener('keydown', handleKeyDown);
-
-
-        return () => {
-            document.removeEventListener('mousedown', handleClickOutside);
-        };
-    });
 
 </script>
 
 <section class="flex justify-center items-center lg:h-screen">
     
-    <input id="rfidInput" autofocus/>
 
     <div 
         class="content grid
